refactor: Update Home.js to display the last 2 projects
This commit updates the Home.js file to modify the logic for displaying projects. Previously, the code sliced the projectsData array to display the first 2 projects. Now, it has been updated to display the last 2 projects by using the slice(-2) method. This change improves the user experience by showcasing the most recent projects at the top of the page.