```text refactor: Update Projects.js to fix navigation prop in Swiper components
The code changes in this commit update the Projects.js file to fix the navigation prop in the Swiper components. Previously, the navigation prop was set to a boolean value, which caused an error. Now, the navigation prop is set to {true} to enable navigation buttons in the Swiper components. This fix improves the user experience by allowing users to navigate between slides in the carousel.