refactor: Update Global.css to adjust text color light
The code changes in this commit update the Global.css file to adjust the value of the --text-color-light variable. Previously, the value was set to #5f5f5f, but it has been changed to #7d7b7b. This modification improves the readability of lighter gray text in the application, enhancing the overall user experience.