refactor: Update index.js to add ChakraProvider and theme
The code changes in this commit update the index.js file to import and use the ChakraProvider component from the '@chakra-ui/react' package. It also imports the theme from the './components/Theme' file and passes it as a prop to the ChakraProvider component. This modification improves the styling and theming capabilities of the application, enhancing the overall user experience.