refactor: Add portfolio website project to ProjectsData.js
This commit adds a new project to the ProjectsData.js file. The project is titled "Mijn Portfolio Website" and serves as a personal portfolio showcasing projects, skills, and professional experience. The website is built using React and styled with Chakra UI. It includes features such as portfolio and CV presentation, project showcase, navigation and routing, responsive design, and contact functionality.

The addition of this project enhances the portfolio by providing a comprehensive overview of the developer's capabilities to potential employers, clients, and other interested parties.

Link to the project: [https://github.com/ALimborgh/Portfolio-Website](https://github.com/ALimborgh/Portfolio-Website)